+++ CHANGELOG.md
@@ -75,3 +75,4 @@ v1.0.6
   
  - Added extra way to retrieve term definition if not set under description (retrieve from annotation)
  - When searching for a detailed item (such as property), return the first 'in defining ontology' or first one if not found
+ - Added PropertyAnnotation parsing